# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  models.TodoItem:
    properties:
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  models.UpdateTodoItemInput:
    properties:
      description:
        type: string
      done:
        type: boolean
      status:
        maximum: 3
        minimum: 0
        type: integer
      title:
        type: string
    required:
    - status
    type: object
host: localhost:8787
info:
  contact: {}
  description: API Server for TodoList Application
  title: Todo App API
  version: "1.0"
paths:
  /api/items:
    get:
      consumes:
      - application/json
      description: get all items
      operationId: get-all-items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TodoItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get All Todo Items
      tags:
      - todo item
    post:
      consumes:
      - application/json
      description: create todo item
      operationId: create-todo-item
      parameters:
      - description: todo item info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TodoItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create Todo Item
      tags:
      - todo item
  /api/items/:id:
    delete:
      consumes:
      - application/json
      description: delete todo item by id
      operationId: delete-todo-item-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete Todo Item By Id
      tags:
      - todo item
    get:
      consumes:
      - application/json
      description: get todo item by id
      operationId: get-todo-item-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Todo Item By Id
      tags:
      - todo item
    put:
      consumes:
      - application/json
      description: update todo item by id
      operationId: update-todo-item-by-id
      parameters:
      - description: todo item update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update Todo Item By Id
      tags:
      - todo item
  /api/items/todo-list-page:
    get:
      consumes:
      - text/html
      description: get all todo items in template
      operationId: get-all-todo-items-in-template
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get All Todo Items In Template
      tags:
      - todo item
swagger: "2.0"